requirements.txt
altair==5.0.1
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.3.3
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
fastjsonschema==2.19.1
Flask==2.2.5
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==7.34.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab_widgets==3.0.10
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.7.0
nest-asyncio==1.6.0
numpy==1.21.6
open3d==0.17.0
packaging==23.2
pandas==1.3.5
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.24.4
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
requests==2.31.0
retrying==1.3.4
rich==13.7.1
six==1.16.0
smmap==5.0.1
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wincertstore==0.2
zipp==3.15.0
